Implement remote communication and remote TTDevice for Blackhole
