+++ device/api/umd/device/tt_device/tt_device.hpp
@@ -242,6 +242,8 @@ public:
 
     ArcTelemetryReader *get_arc_telemetry_reader() const;
 
+    tt_xy_pair get_arc_core() const;
+
     FirmwareInfoProvider *get_firmware_info_provider() const;
 
     virtual uint32_t get_clock() = 0;
@@ -308,6 +310,8 @@ protected:
     ChipInfo chip_info;
 
     bool is_remote_tt_device = false;
+
+    tt_xy_pair arc_core;
 
 private:
     virtual void pre_init_hook(){};
