+++ device/topology/topology_discovery_wormhole.cpp
@@ -232,14 +232,14 @@ std::optional<eth_coord_t> TopologyDiscoveryWormhole::get_remote_eth_coord(Chip*
 }
 
 std::unique_ptr<RemoteChip> TopologyDiscoveryWormhole::create_remote_chip(
-    eth_coord_t eth_coord, Chip* gateway_chip, std::set<uint32_t> gateway_eth_channels) {
+    std::optional<eth_coord_t> eth_coord, Chip* gateway_chip, std::set<uint32_t> gateway_eth_channels) {
     if (is_running_on_6u) {
         return nullptr;
     }
+    eth_coord_t remote_chip_eth_coord = eth_coord.has_value() ? eth_coord.value() : eth_coord_t{0, 0, 0, 0};
 
-    auto local_chip = dynamic_cast<LocalChip*>(gateway_chip);
-
-    return RemoteChip::create(local_chip, eth_coord, gateway_eth_channels, sdesc_path);
+    return RemoteChip::create(
+        dynamic_cast<LocalChip*>(gateway_chip), remote_chip_eth_coord, gateway_eth_channels, sdesc_path);
 }
 
 uint32_t TopologyDiscoveryWormhole::get_remote_eth_channel(Chip* chip, tt_xy_pair local_eth_core) {
