+++ device/tt_device/blackhole_tt_device.cpp
@@ -5,6 +5,9 @@
 
 #include <sys/mman.h>  // for MAP_FAILED
 
+#include <fmt/format.h>
+#include <fmt/ranges.h> 
+#include <iostream>
 #include <tt-logger/tt-logger.hpp>
 
 #include "umd/device/arch/blackhole_implementation.hpp"
@@ -137,6 +140,8 @@ ChipInfo BlackholeTTDevice::get_chip_info() {
 }
 
 void BlackholeTTDevice::wait_arc_core_start(const uint32_t timeout_ms) {
+    std::cout << "wait arc core start " << arc_core.x << " " << arc_core.y << std::endl;
+
     auto start = std::chrono::system_clock::now();
     uint32_t arc_boot_status;
     while (true) {
@@ -150,14 +155,14 @@ void BlackholeTTDevice::wait_arc_core_start(const uint32_t timeout_ms) {
         auto end = std::chrono::system_clock::now();  // End time
         auto duration = std::chrono::duration_cast<std::chrono::milliseconds>(end - start);
         if (duration.count() > timeout_ms) {
-            log_error(
-                LogSiliconDriver,
+            throw std::runtime_error(fmt::format(
                 "Timed out after waiting {} ms for arc core ({}, {}) to start",
                 timeout_ms,
                 arc_core.x,
-                arc_core.y);
+                arc_core.y));
         }
     }
+    std::cout << "arc core started" << std::endl;
 }
 
 uint32_t BlackholeTTDevice::get_clock() {
