+++ device/tt_device/remote_wormhole_tt_device.cpp
@@ -17,11 +17,11 @@ RemoteWormholeTTDevice::RemoteWormholeTTDevice(
 }
 
 void RemoteWormholeTTDevice::read_from_device(void *mem_ptr, tt_xy_pair core, uint64_t addr, uint32_t size) {
-    remote_communication_->read_non_mmio(target_chip_, core, mem_ptr, addr, size);
+    remote_communication_->read_non_mmio(core, mem_ptr, addr, size);
 }
 
 void RemoteWormholeTTDevice::write_to_device(const void *mem_ptr, tt_xy_pair core, uint64_t addr, uint32_t size) {
-    remote_communication_->write_to_non_mmio(target_chip_, core, mem_ptr, addr, size);
+    remote_communication_->write_to_non_mmio(core, mem_ptr, addr, size);
 }
 
 void RemoteWormholeTTDevice::wait_for_non_mmio_flush() { remote_communication_->wait_for_non_mmio_flush(); }
